refactor: simplify assignment workflow structure
